Reduce compilation time, easy selection of tests
To reduce compilation time the tests for each SoftDevice version is merged into one executable.

Selecting tests and tags is done with command line arguments.
