+++ test/softdevice_api/testcase_advertising.cpp
@@ -36,7 +36,6 @@
  */
 
 // Test framework
-#define CATCH_CONFIG_MAIN
 #include "catch2/catch.hpp"
 
 #if NRF_SD_BLE_API >= 6
@@ -55,19 +54,19 @@
 #include <sstream>
 #include <thread>
 
-using namespace testutil;
-
-// Indicates if an error has occurred in a callback.
-// The test framework is not thread safe so this variable is used to communicate that an issues has
-// occurred in a callback.
-bool error = false;
-
-TEST_CASE("test_advertising_api")
+TEST_CASE("test_advertising_api", "[advertising][known_error]")
 {
+    using namespace testutil;
+
     auto env = ::test::getEnvironment();
     REQUIRE(env.serialPorts.size() >= 2);
     const auto central    = env.serialPorts.at(0);
     const auto peripheral = env.serialPorts.at(1);
+
+    // Indicates if an error has occurred in a callback.
+    // The test framework is not thread safe so this variable is used to communicate that an issues
+    // has occurred in a callback.
+    bool error = false;
 
 #if NRF_SD_BLE_API == 6
     SECTION("test_extended")
