+++ test/softdevice_api/testcase_issue_gh_112.cpp
@@ -36,11 +36,13 @@
  */
 
 // Test framework
-#define CATCH_CONFIG_MAIN
 #include "catch2/catch.hpp"
 
+// This issue is only relevant for SoftDevice API >= 3
+#if NRF_SD_BLE_API >= 3
+
 // Logging support
-#include <internal/log.h>
+#include "internal/log.h"
 
 // Test support
 #include <test_setup.h>
@@ -53,119 +55,99 @@
 #include <string>
 #include <thread>
 
-constexpr uint16_t BLE_UUID_HEART_RATE_SERVICE          = 0x180D;
-constexpr uint16_t BLE_UUID_HEART_RATE_MEASUREMENT_CHAR = 0x2A37;
-constexpr uint16_t BLE_UUID_CCCD                        = 0x2902;
-
-// Indicates if an error has occurred in a callback.
-// The test framework is not thread safe so this variable is used to communicate that an issues has
-// occurred in a callback.
-bool error = false;
-
-// Set to true when the test is complete
-bool testComplete = false;
-
-enum AuthenticationType {
-    LEGACY_PASSKEY,
-    LEGACY_OOB,
-};
-
-AuthenticationType authType;
-
-uint32_t setupPeripheral(const std::shared_ptr<testutil::AdapterWrapper> &p,
-                         const std::string &advertisingName,
-                         const std::vector<uint8_t> &initialCharacteristicValue,
-                         const uint16_t characteristicValueMaxLength)
-{
-    // Setup the advertisement data
-    std::vector<uint8_t> advertisingData;
-    testutil::appendAdvertisingName(advertisingData, advertisingName);
-    advertisingData.push_back(3); // Length of upcoming advertisement type
-    advertisingData.push_back(BLE_GAP_AD_TYPE_16BIT_SERVICE_UUID_COMPLETE);
-
-    // Store BLE_UUID_HEART_RATE_SERVICE in little-endian format.
-    advertisingData.push_back(BLE_UUID_HEART_RATE_SERVICE & 0xFF);
-    advertisingData.push_back((BLE_UUID_HEART_RATE_SERVICE & 0xFF00) >> 8);
-
-    auto err_code = p->setupAdvertising(advertisingData);
-    if (err_code != NRF_SUCCESS)
-    {
-        NRF_LOG(p->role() << " Error setting advertising data, "
-                          << ", " << testutil::errorToString(err_code));
-        return err_code;
-    }
-
-    // Setup service, use service UUID specified in scratchpad.target_service
-    err_code =
-        sd_ble_gatts_service_add(p->unwrap(), BLE_GATTS_SRVC_TYPE_PRIMARY,
-                                 &(p->scratchpad.target_service), &(p->scratchpad.service_handle));
-    if (err_code != NRF_SUCCESS)
-    {
-        NRF_LOG(p->role() << " Error adding GATTS service, "
-                          << ", " << testutil::errorToString(err_code));
-        return err_code;
-    }
-
-    // Setup characteristic, use characteristic UUID specified in scratchpad.target_characteristic
-    ble_gatts_char_md_t char_md;
-    ble_gatts_attr_md_t cccd_md;
-    ble_gatts_attr_t attr_char_value;
-    ble_gatts_attr_md_t attr_md;
-
-    memset(&cccd_md, 0, sizeof(cccd_md));
-
-    BLE_GAP_CONN_SEC_MODE_SET_OPEN(&cccd_md.read_perm);
-    BLE_GAP_CONN_SEC_MODE_SET_OPEN(&cccd_md.write_perm);
-    cccd_md.vloc = BLE_GATTS_VLOC_STACK;
-
-    memset(&char_md, 0, sizeof(char_md));
-
-    char_md.char_props.read          = 1;
-    char_md.char_props.notify        = 1;
-    char_md.char_props.write         = 1;
-    char_md.char_props.write_wo_resp = 1;
-    char_md.p_char_user_desc         = nullptr;
-    char_md.p_char_pf                = nullptr;
-    char_md.p_user_desc_md           = nullptr;
-    char_md.p_cccd_md                = &cccd_md;
-    char_md.p_sccd_md                = nullptr;
-
-    memset(&attr_md, 0, sizeof(attr_md));
-
-    BLE_GAP_CONN_SEC_MODE_SET_OPEN(&attr_md.read_perm);
-    BLE_GAP_CONN_SEC_MODE_SET_OPEN(&attr_md.write_perm);
-    attr_md.vloc    = BLE_GATTS_VLOC_STACK;
-    attr_md.rd_auth = 0;
-    attr_md.wr_auth = 0;
-    attr_md.vlen    = 1;
-
-    memset(&attr_char_value, 0, sizeof(attr_char_value));
-
-    attr_char_value.p_uuid    = &(p->scratchpad.target_characteristic);
-    attr_char_value.p_attr_md = &attr_md;
-    attr_char_value.init_len  = static_cast<uint16_t>(initialCharacteristicValue.size());
-    attr_char_value.init_offs = 0;
-    attr_char_value.max_len   = characteristicValueMaxLength;
-    attr_char_value.p_value   = const_cast<uint8_t *>(initialCharacteristicValue.data());
-
-    err_code = sd_ble_gatts_characteristic_add(p->unwrap(), p->scratchpad.service_handle, &char_md,
-                                               &attr_char_value,
-                                               &(p->scratchpad.gatts_characteristic_handle));
-
-    NRF_LOG(p->role() << " Error adding GATTS characteristic"
-                      << ", " << testutil::errorToString(err_code));
-
-    return err_code;
-}
-
-TEST_CASE("test_security")
+TEST_CASE("test_issue_gh_112")
 {
     auto env = ::test::getEnvironment();
     REQUIRE(env.serialPorts.size() >= 2);
     const auto central    = env.serialPorts.at(0);
     const auto peripheral = env.serialPorts.at(1);
 
-    SECTION("legacy_passkey")
+    constexpr uint16_t BLE_UUID_HEART_RATE_SERVICE          = 0x180D;
+    constexpr uint16_t BLE_UUID_HEART_RATE_MEASUREMENT_CHAR = 0x2A37;
+    constexpr uint16_t BLE_UUID_CCCD                        = 0x2902;
+
+    // Indicates if an error has occurred in a callback.
+    // The test framework is not thread safe so this variable is used to communicate that an
+    // issues has occurred in a callback.
+    bool error = false;
+
+    // Set to true when the test is complete
+    bool testComplete = false;
+
+    auto setupPeripheral = [&](const std::shared_ptr<testutil::AdapterWrapper> &p,
+                               const std::string &advertisingName,
+                               const std::vector<uint8_t> &initialCharaciteristicValue,
+                               const uint16_t characteristicValueMaxLength) -> uint32_t {
+        // Setup the advertisement data
+        std::vector<uint8_t> advertisingData;
+        testutil::appendAdvertisingName(advertisingData, advertisingName);
+        advertisingData.push_back(3); // Length of upcoming advertisement type
+        advertisingData.push_back(BLE_GAP_AD_TYPE_16BIT_SERVICE_UUID_COMPLETE);
+
+        // Store BLE_UUID_HEART_RATE_SERVICE in little - endian format.
+        advertisingData.push_back(BLE_UUID_HEART_RATE_SERVICE & 0xFF);
+        advertisingData.push_back((BLE_UUID_HEART_RATE_SERVICE & 0xFF00) >> 8);
+
+        auto err_code = p->setupAdvertising(advertisingData);
+        if (err_code != NRF_SUCCESS)
+            return err_code;
+
+        // Setup service, use service UUID specified in scratchpad.target_service
+        err_code = sd_ble_gatts_service_add(p->unwrap(), BLE_GATTS_SRVC_TYPE_PRIMARY,
+                                            &(p->scratchpad.target_service),
+                                            &(p->scratchpad.service_handle));
+        if (err_code != NRF_SUCCESS)
+            return err_code;
+
+        // Setup characteristic, use characteristic UUID specified in
+        // scratchpad.target_characteristic
+        ble_gatts_char_md_t char_md;
+        ble_gatts_attr_md_t cccd_md;
+        ble_gatts_attr_t attr_char_value;
+        ble_gatts_attr_md_t attr_md;
+
+        memset(&cccd_md, 0, sizeof(cccd_md));
+
+        BLE_GAP_CONN_SEC_MODE_SET_OPEN(&cccd_md.read_perm);
+        BLE_GAP_CONN_SEC_MODE_SET_OPEN(&cccd_md.write_perm);
+        cccd_md.vloc = BLE_GATTS_VLOC_STACK;
+
+        memset(&char_md, 0, sizeof(char_md));
+
+        char_md.char_props.read          = 1;
+        char_md.char_props.notify        = 1;
+        char_md.char_props.write         = 1;
+        char_md.char_props.write_wo_resp = 1;
+        char_md.p_char_user_desc         = nullptr;
+        char_md.p_char_pf                = nullptr;
+        char_md.p_user_desc_md           = nullptr;
+        char_md.p_cccd_md                = &cccd_md;
+        char_md.p_sccd_md                = nullptr;
+
+        memset(&attr_md, 0, sizeof(attr_md));
+
+        BLE_GAP_CONN_SEC_MODE_SET_OPEN(&attr_md.read_perm);
+        BLE_GAP_CONN_SEC_MODE_SET_OPEN(&attr_md.write_perm);
+        attr_md.vloc    = BLE_GATTS_VLOC_STACK;
+        attr_md.rd_auth = 0;
+        attr_md.wr_auth = 0;
+        attr_md.vlen    = 1;
+
+        memset(&attr_char_value, 0, sizeof(attr_char_value));
+
+        attr_char_value.p_uuid    = &(p->scratchpad.target_characteristic);
+        attr_char_value.p_attr_md = &attr_md;
+        attr_char_value.init_len  = static_cast<uint16_t>(initialCharaciteristicValue.size());
+        attr_char_value.init_offs = 0;
+        attr_char_value.max_len   = characteristicValueMaxLength;
+        attr_char_value.p_value   = const_cast<uint8_t *>(initialCharaciteristicValue.data());
+
+        return sd_ble_gatts_characteristic_add(p->unwrap(), p->scratchpad.service_handle, &char_md,
+                                               &attr_char_value,
+                                               &(p->scratchpad.gatts_characteristic_handle));
+    };
+
+    SECTION("reproduce_error")
     {
         const auto baudRate = central.baudRate;
 
@@ -176,12 +158,12 @@ TEST_CASE("test_security")
         const auto peripheralAdvName = "peripheral";
 
         // Instantiate an adapter to use as BLE Central in the test
-        auto c =
-            std::make_shared<testutil::AdapterWrapper>(testutil::Central, central.port, baudRate);
+        auto c = std::make_shared<testutil::AdapterWrapper>(testutil::Central, central.port,
+                                                            baudRate, 150);
 
         // Instantiated an adapter to use as BLE Peripheral in the test
         auto p = std::make_shared<testutil::AdapterWrapper>(testutil::Peripheral, peripheral.port,
-                                                            baudRate);
+                                                            baudRate, 150);
 
         // Use Heart rate service and characteristics as target for testing but
         // the values sent are not according to the Heart Rate service
@@ -195,6 +177,7 @@ TEST_CASE("test_security")
 
         c->scratchpad.target_descriptor.uuid = BLE_UUID_CCCD;
         c->scratchpad.target_descriptor.type = BLE_UUID_TYPE_BLE;
+        c->scratchpad.mtu                    = 150;
 
         // BLE Peripheral scratchpad
         p->scratchpad.target_service.uuid = BLE_UUID_HEART_RATE_SERVICE;
@@ -205,26 +188,18 @@ TEST_CASE("test_security")
 
         p->scratchpad.target_descriptor.uuid = BLE_UUID_CCCD;
         p->scratchpad.target_descriptor.type = BLE_UUID_TYPE_BLE;
+        p->scratchpad.mtu                    = 150;
 
         REQUIRE(sd_rpc_log_handler_severity_filter_set(c->unwrap(), env.driverLogLevel) ==
                 NRF_SUCCESS);
         REQUIRE(sd_rpc_log_handler_severity_filter_set(p->unwrap(), env.driverLogLevel) ==
                 NRF_SUCCESS);
 
-        c->setGapEventCallback([&c, &p, peripheralAdvName](const uint16_t eventId,
-                                                           const ble_gap_evt_t *gapEvent) -> bool {
+        c->setGapEventCallback([&](const uint16_t eventId, const ble_gap_evt_t *gapEvent) -> bool {
             switch (eventId)
             {
                 case BLE_GAP_EVT_CONNECTED:
-                {
-                    authType            = LEGACY_PASSKEY;
-                    const auto err_code = c->startAuthentication(true, true, false, true);
-
-                    if (err_code != NRF_SUCCESS)
-                    {
-                        error = true;
-                    }
-                }
+                    c->startServiceDiscovery(BLE_UUID_TYPE_BLE, BLE_UUID_HEART_RATE_SERVICE);
                     return true;
                 case BLE_GAP_EVT_DISCONNECTED:
                     return true;
@@ -270,7 +245,6 @@ TEST_CASE("test_security")
                     const auto err_code = sd_ble_gap_conn_param_update(
                         c->unwrap(), c->scratchpad.connection_handle,
                         &(gapEvent->params.conn_param_update_request.conn_params));
-
                     if (err_code != NRF_SUCCESS)
                     {
                         NRF_LOG(c->role() << " Conn params update failed, err_code " << err_code);
@@ -278,87 +252,18 @@ TEST_CASE("test_security")
                     }
                 }
                     return true;
-                case BLE_GAP_EVT_SEC_PARAMS_REQUEST:
-                {
-                    ble_gap_sec_keyset_t keyset;
-                    memset(&keyset, 0, sizeof(ble_gap_sec_keyset_t));
-
-                    ble_gap_sec_keys_t keys_own;
-                    ble_gap_sec_keys_t keys_peer;
-                    memset(&keys_own, 0, sizeof(keys_own));
-                    memset(&keys_peer, 0, sizeof(keys_peer));
-
-                    keyset.keys_own  = keys_own;
-                    keyset.keys_peer = keys_peer;
-
-                    const auto err_code = c->securityParamsReply(keyset);
-
-                    if (err_code != NRF_SUCCESS)
-                    {
-                        error = true;
-                    }
-                }
-                    return true;
-                case BLE_GAP_EVT_PASSKEY_DISPLAY:
-                {
-                    size_t size = 0;
-
-                    switch (p->scratchpad.key_type)
-                    {
-                        case BLE_GAP_AUTH_KEY_TYPE_PASSKEY:
-                            size = 6;
-                            break;
-                        case BLE_GAP_AUTH_KEY_TYPE_OOB:
-                            size = 16;
-                            break;
-                        default:
-                            break;
-                    }
-
-                    if (size != 0)
-                    {
-                        std::memcpy(c->scratchpad.key, gapEvent->params.passkey_display.passkey,
-                                    size);
-                    }
-
-                    const auto err_code =
-                        p->authKeyReply(p->scratchpad.key_type, c->scratchpad.key);
-
-                    if (err_code != NRF_SUCCESS)
-                    {
-                        error = true;
-                    }
-                }
-                    return true;
-                case BLE_GAP_EVT_SEC_REQUEST:
-                    return true;
-                case BLE_GAP_EVT_AUTH_KEY_REQUEST:
-                    return true;
-                case BLE_GAP_EVT_CONN_SEC_UPDATE:
-                    return true;
-                case BLE_GAP_EVT_AUTH_STATUS:
-                    if (gapEvent->params.auth_status.auth_status == BLE_GAP_SEC_STATUS_SUCCESS)
-                    {
-                        testComplete = true;
-                    }
-                    else
-                    {
-                        error = true;
-                    }
-
-                    return true;
                 default:
                     return false;
             }
         });
 
-        c->setGattcEventCallback([&c](const uint16_t eventId,
-                                      const ble_gattc_evt_t *gattcEvent) -> bool {
+        c->setGattcEventCallback([&](const uint16_t eventId,
+                                     const ble_gattc_evt_t *gattcEvent) -> bool {
             switch (eventId)
             {
                 case BLE_GATTC_EVT_PRIM_SRVC_DISC_RSP:
                 {
-                    NRF_LOG(c->role() << " BLE_GATTC_EVT_PRIM_SRVC_DISC_RSP");
+                    NRF_LOG(c->role() << " Received service discovery response.");
 
                     if (gattcEvent->gatt_status != NRF_SUCCESS)
                     {
@@ -490,6 +395,17 @@ TEST_CASE("test_security")
                     NRF_LOG(c->role()
                             << " Received descriptor discovery response, descriptor count: "
                             << count);
+
+                    // Change the MTU
+                    const auto err_code = sd_ble_gattc_exchange_mtu_request(
+                        c->unwrap(), c->scratchpad.connection_handle, c->scratchpad.mtu);
+
+                    if (err_code != NRF_SUCCESS)
+                    {
+                        NRF_LOG(c->role()
+                                << " MTU exchange request failed, err_code: " << err_code);
+                        error = true;
+                    }
                 }
                     return true;
                 case BLE_GATTC_EVT_WRITE_RSP:
@@ -501,22 +417,63 @@ TEST_CASE("test_security")
                                           << testutil::gattStatusToString(gattcEvent->gatt_status));
                         error = true;
                     }
+                    else
+                    {
+                        testComplete = true;
+                    }
 
-                    testComplete = true;
+                    return true;
+                case BLE_GATTC_EVT_EXCHANGE_MTU_RSP:
+                {
+                    auto const server_rx_mtu = gattcEvent->params.exchange_mtu_rsp.server_rx_mtu;
+                    NRF_LOG(c->role()
+                            << " MTU response received. New ATT_MTU is " << server_rx_mtu);
 
+                    // Write some data to characteristic to trigger reported error
+                    std::vector<uint8_t> data;
+
+                    // Send negotiated MTU - 3 bytes  used by GATT header
+                    testutil::appendRandomData(data, c->scratchpad.mtu - 3);
+
+                    const auto err_code = c->writeCharacteristicValue(
+                        c->scratchpad.characteristic_value_handle, data);
+
+                    if (err_code != NRF_SUCCESS)
+                    {
+                        NRF_LOG(c->role()
+                                << " Error writing data to characteristic. err_code: " << err_code);
+                        error = true;
+                    }
+                }
                     return true;
                 default:
                     return false;
             }
         });
 
-        c->setEventCallback([&c](const ble_evt_t *p_ble_evt) -> bool {
-            const auto eventId = p_ble_evt->header.evt_id;
-            NRF_LOG(c->role() << " Received an un-handled event with ID: " << eventId);
-            return true;
-        });
+        c->setGattsEventCallback(
+            [&](const uint16_t eventId, const ble_gatts_evt_t *gattsEvent) -> bool {
+                switch (eventId)
+                {
+                    case BLE_GATTS_EVT_EXCHANGE_MTU_REQUEST:
+                    {
+                        const auto err_code = sd_ble_gatts_exchange_mtu_reply(
+                            c->unwrap(), c->scratchpad.connection_handle, c->scratchpad.mtu);
 
-        p->setGapEventCallback([&p](const uint16_t eventId, const ble_gap_evt_t *gapEvent) {
+                        if (err_code != NRF_SUCCESS)
+                        {
+                            NRF_LOG(c->role() << " MTU exchange request reply failed, err_code: "
+                                              << err_code);
+                            error = true;
+                        }
+                    }
+                        return true;
+                    default:
+                        return false;
+                }
+            });
+
+        p->setGapEventCallback([&](const uint16_t eventId, const ble_gap_evt_t *gapEvent) {
             switch (eventId)
             {
                 case BLE_GAP_EVT_CONNECTED:
@@ -525,9 +482,11 @@ TEST_CASE("test_security")
                 {
                     // Use scratchpad defaults when advertising
                     NRF_LOG(p->role() << " Starting advertising.");
+
                     const auto err_code = p->startAdvertising();
                     if (err_code != NRF_SUCCESS)
                     {
+                        NRF_LOG(p->role() << " Error starting advertising after disconnect.");
                         error = true;
                     }
                 }
@@ -536,42 +495,20 @@ TEST_CASE("test_security")
                     return true;
                 case BLE_GAP_EVT_SEC_PARAMS_REQUEST:
                 {
-                    ble_gap_sec_keyset_t keyset;
-                    std::memset(&keyset, 0, sizeof(keyset));
-
                     const auto err_code =
-                        p->securityParamsReply(BLE_GAP_SEC_STATUS_SUCCESS, keyset);
+                        sd_ble_gap_sec_params_reply(p->unwrap(), p->scratchpad.connection_handle,
+                                                    BLE_GAP_SEC_STATUS_SUCCESS, nullptr, nullptr);
 
                     if (err_code != NRF_SUCCESS)
                     {
                         error = true;
+                        NRF_LOG(p->role() << "Failed reply with GAP security parameters. "
+                                          << testutil::errorToString(err_code));
                     }
                 }
                     return true;
-                case BLE_GAP_EVT_PASSKEY_DISPLAY:
-                    return true;
-                case BLE_GAP_EVT_SEC_REQUEST:
-                    return true;
-                case BLE_GAP_EVT_AUTH_KEY_REQUEST:
-                    return true;
-                case BLE_GAP_EVT_CONN_SEC_UPDATE:
-                    return true;
-                case BLE_GAP_EVT_AUTH_STATUS:
-                    if (gapEvent->params.auth_status.auth_status == BLE_GAP_SEC_STATUS_SUCCESS)
-                    {
-                        // Move on to the next type of bonding/pairing
-                        testComplete = true;
-                    }
-                    else
-                    {
-                        error = true;
-                    }
-
-                    return true;
                 case BLE_GATTS_EVT_SYS_ATTR_MISSING:
                 {
-                    NRF_LOG(p->role() << " BLE_GATTS_EVT_SYS_ATTR_MISSING");
-
                     const auto err_code = sd_ble_gatts_sys_attr_set(
                         p->unwrap(), p->scratchpad.connection_handle, nullptr, 0, 0);
 
@@ -588,11 +525,27 @@ TEST_CASE("test_security")
             }
         });
 
-        p->setEventCallback([&p](const ble_evt_t *p_ble_evt) -> bool {
-            const auto eventId = p_ble_evt->header.evt_id;
-            NRF_LOG(p->role() << " Received an un-handled event with ID: " << eventId);
-            return true;
-        });
+        p->setGattsEventCallback(
+            [&](const uint16_t eventId, const ble_gatts_evt_t *gattsEvent) -> bool {
+                switch (eventId)
+                {
+                    case BLE_GATTS_EVT_EXCHANGE_MTU_REQUEST:
+                    {
+                        const auto err_code = sd_ble_gatts_exchange_mtu_reply(
+                            p->unwrap(), p->scratchpad.connection_handle, p->scratchpad.mtu);
+
+                        if (err_code != NRF_SUCCESS)
+                        {
+                            error = true;
+                            NRF_LOG(p->role() << " MTU exchange request reply failed. "
+                                              << testutil::errorToString(err_code));
+                        }
+                    }
+                        return true;
+                    default:
+                        return false;
+                }
+            });
 
         // Open the adapters
         REQUIRE(c->open() == NRF_SUCCESS);
@@ -601,14 +554,14 @@ TEST_CASE("test_security")
         REQUIRE(c->configure() == NRF_SUCCESS);
         REQUIRE(p->configure() == NRF_SUCCESS);
 
-        // Starting the scan starts the sequence of operations to get a connection established
-        REQUIRE(c->startScan() == NRF_SUCCESS);
-
         REQUIRE(setupPeripheral(p, peripheralAdvName, {0x00}, p->scratchpad.mtu) == NRF_SUCCESS);
         REQUIRE(p->startAdvertising() == NRF_SUCCESS);
 
+        // Starting the scan starts the sequence of operations to get a connection established
+        REQUIRE(c->startScan() == NRF_SUCCESS);
+
         // Wait for the test to complete
-        std::this_thread::sleep_for(std::chrono::seconds(5));
+        std::this_thread::sleep_for(std::chrono::seconds(3));
 
         REQUIRE(error == false);
         REQUIRE(testComplete == true);
@@ -620,3 +573,11 @@ TEST_CASE("test_security")
         sd_rpc_adapter_delete(p->unwrap());
     }
 }
+
+#else
+TEST_CASE("test_issue_gh_112")
+{
+    INFO("Not relevant for SoftDevice API version < 3")
+}
+
+#endif // This issue is only relevant for SoftDevice API >= 3
