+++ test/transport/test_uart_boost.cpp
@@ -36,6 +36,7 @@
  */
 
 // Logging support
+#define NRF_LOG_SETUP
 #include "internal/log.h"
 
 // Test framework
