+++ test/util/include/test_util_role.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include "ble.h"
+#include <iostream>
+
+namespace testutil {
+enum Role { Central = BLE_GAP_ROLE_CENTRAL, Peripheral = BLE_GAP_ROLE_PERIPH };
+
+std::ostream &operator<<(std::ostream &s, const Role &role);
+} //  namespace testutil
